zephyr: Add support for CONFIG_FPU
Allow this to be enabled for Zephyr. Since we are using the minimal C
library we need to define our own version of a few functions. Add these
for Cortex-M4, taken from the existing ECOS code.

It happens that Zephyr uses CONFIG_FPU, the same as ECOS, so there is
nothing needed in config_chip.h for this case. If CONFIG_FPU is enabled
on a new platform, some work will be needed to make it build.

BUG=b:180023514
BRANCH=none
TEST=build zephyr volteer with CONFIG_NEWLIB_LIBC and see no warnings
Build for volteer

Signed-off-by: Simon Glass <sjg@chromium.org>
Change-Id: I91187560b7e699b7f77ccfa5990617606f7fd6fa
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/2691414
Reviewed-by: Keith Short <keithshort@chromium.org>