cbi: Add CONFIG_BYPASS_CBI_EEPROM_WP_CHECK
In some projects, we modify the cbi eeprom write protection
via the hardware design. So, the gpio pin GPIO_WP is
decoupled from the cbi eeprom write protection status. With
this change, we can update the cbi eeprom data on the fly
for the dogfood device during the development.
In order to bypass the cbi eeprom write protection check,
we add a CONFIG_BYPASS_CBI_EEPROM_WP_CHECK to skip the
gpio value.
Since this modification is only for the developing device,
we add the CONFIG_SYSTEM_UNLOCKED check to avoid someone
mis-use this config.

BUG=b:169034911
BRANCH=volteer
TEST=`ectool cbi set` is working in drobit/copano

Change-Id: I3f35159883f75965b1009c45c25ac2cd1c6e7c0d
Signed-off-by: Zhuohao Lee <zhuohao@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/2719724
Reviewed-by: Aseda Aboagye <aaboagye@chromium.org>