bq25720: Add VSYS_TH2 tuning support
This adds support for tuning the bq25720 VSYS_TH2 threshold so that a
custom threshold voltage can be specified. The default 3.2v for 1S
batteries and 5.9v for 2S (or more) batteries may need to be adjusted to
improve system stability. When VSYS droops to this threshold, PROCHOT is
asserted to reduce system power demand.

BRANCH=none
BUG=b:183474163
TEST=with rest of patch stack, booted brya to the OS without a battery

Signed-off-by: Caveh Jalali <caveh@chromium.org>
Change-Id: I7f52a369371eb93860b14ad8438b790af231c450
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/2798979
Reviewed-by: Keith Short <keithshort@chromium.org>
Commit-Queue: Keith Short <keithshort@chromium.org>