driver: icm: fix use of I2C_PORT_ACCEL
The ICM driver is currently using I2C_PORT_ACCEL incorrectly as a
CONFIG_ value.

Update the use cases to a new config option that selects between SPI
and I2C communication specifically for the chip. To avoid a lot of
device.h changes, the value of the config value is automatically
inferred if not explicitly set.

BRANCH=none
BUG=b:185392974, b:146065507
TEST=zmake testall
TEST=make buildall

Signed-off-by: Yuval Peress <peress@chromium.org>
Change-Id: I8b1a507a76031c2bb1aaf4ca7b14b92252a941f8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/2826920
Commit-Queue: Keith Short <keithshort@chromium.org>
Reviewed-by: Keith Short <keithshort@chromium.org>