ectool: Implement I2C speed control command
This adds the new "i2cspeed" command. The command takes an I2C port
(bus) number as an argument. Without a 2nd argument, the current I2C
speed is reported. If a 2nd argument is given, the I2C bus speed
accordingly, the unit is assumed to be Kbps. In practice, only 3 speeds
are supported: 100, 400, 1000 kHz.

BRANCH=none
BUG=b:201039003
TEST=with follow-on patches, switched I2C bus speed between 400 kHz
	and 1 MHz using ectool.

Change-Id: I80cd436c85c368f13357a4a80c3179e1fb483cb1
Signed-off-by: Caveh Jalali <caveh@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/3181508
Reviewed-by: Daisuke Nojiri <dnojiri@chromium.org>
Reviewed-by: Boris Mittelberg <bmbm@google.com>