test: move spi_flash_reg to its own module
Move the logic of the spi_flash_reg.c from common/ to a separate
directory. Also, move the tests. Update the build.mk system to still
build this in the same manner as well as CMake to use it as a library.

BRANCH=none
BUG=b:246839061
TEST=make BOARD=dooly
TEST=zmake build hoglin
TEST=./twister -T common

Signed-off-by: Yuval Peress <peress@google.com>
Change-Id: Icc6b4fce1619c2c6a86c040f30390b7788171a9d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/3935025
Reviewed-by: Keith Short <keithshort@chromium.org>
Reviewed-by: Aaron Massey <aaronmassey@google.com>
Code-Coverage: Zoss <zoss-cl-coverage@prod.google.com>